+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## master (unreleased)
 
-+### Bug fixes
+### Bug fixes
 
 * Fix `Style/RedundantParentheses` with hash literal as first argument to `super`. ([@maxh][])
+* [#6086](https://github.com/bbatsov/rubocop/issues/6086): Fix an error for `Gemspec/OrderedDependencies` when using method call to gem names in gemspec. ([@koic][])
 
 ## 0.58.0 (2018-07-07)